Revert "working on transfering register allocation and lifetime analysis to the abstract machine."
This reverts commit 68da1674ac8ecfc4000c42364d8d1cfd553a55e2.

+++ exp/include/imr/block.h
@@ -24,8 +24,8 @@
  *
  */
 typedef struct Block {
-    u64 length;
-    u64 capacity;
+    u32 length;
+    u32 capacity;
     Instruction *buffer;
 } Block;
